Cria a estrutura de pastas
