+++ app/src/main/java/com/app/criptofavorite/view/ui/notifications/NotificationsViewModel.kt
@@ -1,13 +1,13 @@
-package com.app.criptofavorite.ui.home
+package com.app.criptofavorite.view.ui.notifications
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class NotificationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
